refactor: 디테일 페이지 컴포넌트 추상화 레벨 변경
