+++ README.md
@@ -11,3 +11,12 @@
 - [x] 없는 상품의 경우 메인 페이지로 연결되도록 한다.
 - [x] 나에게 선물하기 버튼 클릭 시 로그인이 되어있지 않다면 로그인 페이지로 이동한다.
 - [x] React Hook Form 등의 라이브러리를 사용하지 않으며 React의 form, ref, state만 사용하여 구현한다.
+
+## 2단계 요구사항
+
+- [ ] 상품 상세 페이지에서 상품의 개수를 option API의 giftOrderLimit을 초과한 경우 선택이 불가하게 해요.
+- [ ] 결제 페이지의 Form을 validation 해요. (4주차 미션 결과물 참고)
+  - [ ] 카드 메시지를 입력하지 않으면 메시지를 입력하라고 안내해요.
+  - [ ] 카드 메시지가 100글자가 넘어가면 100자 이내로 입력하라고 안내해요.
+  - [ ] 현금 영수증 checkbox 클릭 시 현금영수증 번호가 입력되었는지 확인해요.
+  - [ ] 현금 영수증 입력은 숫자만 입력하도록 안내해요.
