+++ src/pages/ProductDetail/index.tsx
@@ -1,6 +1,8 @@
+import { Stack } from '@chakra-ui/react';
 import { useNavigate, useParams } from 'react-router-dom';
 
 import { Container } from '@/components/common/layouts/Container';
+import ProductCountSection from '@/components/features/Detail/ProductCountSection';
 import ProductDetailSection from '@/components/features/Detail/ProductDetailSection';
 
 const ProductDetialPage = () => {
@@ -15,7 +17,10 @@ const ProductDetialPage = () => {
 
   return (
     <Container>
-      <ProductDetailSection productId={productId} />
+      <Stack direction="row">
+        <ProductDetailSection productId={productId} />
+        <ProductCountSection productId={productId} />
+      </Stack>
     </Container>
   );
 };
